объединил модули modal и big-picture
