+++ js/modal.js
@@ -1,9 +1,13 @@
 import { isEscKey } from './utils';
-import { renderBigPicture } from './big-picture';
+import { renderComments } from './comments';
 
 const bodyElement = document.body;
 const picturesElement = bodyElement.querySelector('.pictures');
 const bigPictureElement = bodyElement.querySelector('.big-picture');
+const imgElement = bigPictureElement.querySelector('.big-picture__img img');
+const socialCaptionElement = bigPictureElement.querySelector('.social__caption');
+const likesCountElement = bigPictureElement.querySelector('.likes-count');
+const commentTotalElement = bigPictureElement.querySelector('.social__comment-total-count');
 
 const closeModal = () => {
   bodyElement.classList.remove('modal-open');
@@ -22,6 +26,15 @@ const onBigPictureElementClick = (evt) => {
   }
 };
 
+const renderPhotos = (photos, index) => {
+  const { url, description, likes, comments } = photos[index];
+  imgElement.src = url;
+  socialCaptionElement.textContent = description;
+  likesCountElement.textContent = likes;
+  commentTotalElement.textContent = comments.length;
+  renderComments(comments);
+};
+
 bigPictureElement.addEventListener('click', onBigPictureElementClick);
 document.addEventListener('keydown', onDocumentKeydown);
 
@@ -37,7 +50,7 @@ const renderModal = (data) => {
       bigPictureElement.classList.remove('hidden');
       bodyElement.classList.add('modal-open');
 
-      renderBigPicture(data, id);
+      renderPhotos(data, id);
     }
   };
